Refactor the button of Add and Edit Meeting
